+++ core/src/test/java/com/bc/fiduceo/reader/airs/ForReaderFactoryTest_AIRS_L1B_Reader.java
@@ -1,19 +1,23 @@
 /*
- * $Id$
- *
- * Copyright (C) 2010 by Brockmann Consult (info@brockmann-consult.de)
+ * Copyright (C) 2018 Brockmann Consult GmbH
+ * This code was developed for the EC project "Fidelity and Uncertainty in
+ * Climate Data Records from Earth Observations (FIDUCEO)".
+ * Grant Agreement: 638822
  *
  * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License as published by the
- * Free Software Foundation. This program is distributed in the hope it will
- * be useful, but WITHOUT ANY WARRANTY; without even the implied warranty
- * of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU General Public License for more details.
+ * under the terms of the GNU General Public License as published by the Free
+ * Software Foundation; either version 3 of the License, or (at your option)
+ * any later version.
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+ * more details.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+ * A copy of the GNU General Public License should have been supplied along
+ * with this program; if not, see http://www.gnu.org/licenses/
+ *
  */
+
 package com.bc.fiduceo.reader.airs;
 
 import static org.hamcrest.Matchers.*;
@@ -22,6 +26,7 @@ import static org.junit.Assert.*;
 import com.bc.fiduceo.reader.Reader;
 
 public class ForReaderFactoryTest_AIRS_L1B_Reader {
+
     public static void checkInstance(Reader reader) {
         assertThat(reader, is(instanceOf(AIRS_L1B_Reader.class)));
     }