+++ ingestion-tool/src/test/java/com/bc/fiduceo/ingest/IngestionToolTest.java
@@ -59,7 +59,7 @@ public class IngestionToolTest {
 
         ingestionTool.printUsageTo(outputStream);
 
-        assertEquals("ingestion-tool version 1.3.9" + ls +
+        assertEquals("ingestion-tool version 1.4.0-SNAPSHOT" + ls +
                 ls +
                 "usage: ingestion-tool <options>" + ls +
                 "Valid options are:" + ls +